Move more of sysroot's sadness into iPhonePrivate.
# Conflicts:
#	MobileCydia.mm
#	Scripts/Legacy/sysroot.sh
#	Source/Legacy/CyteKit/WebView.h
#	Source/Legacy/CyteKit/WebViewController.mm
#	Source/Legacy/iPhonePrivate.h

Comitter's note: Also used this opportunity to remove most warnings.

+++ External/usr/include/apt-pkg/sourcelist.h
@@ -105,7 +105,7 @@ class pkgSourceList
    // List accessors
    inline const_iterator begin() const {return SrcList.begin();};
    inline const_iterator end() const {return SrcList.end();};
-   inline unsigned int size() const {return SrcList.size();};
+   inline unsigned int size() const {return (unsigned int)SrcList.size();};
    inline bool empty() const {return SrcList.empty();};
 
    bool FindIndex(pkgCache::PkgFileIterator File,
