+++ Source/Models/Packages/Package/Package.h
@@ -174,7 +174,7 @@ static inline uint32_t PackageChangesRadix(Package *self, void *) {
         value.bits.ignored = [self ignored] ? 0 : 1;
         value.bits.upgradable = 1;
     } else {
-        value.bits.timestamp = [self seen] >> 2;
+        value.bits.timestamp = (uint32_t)([self seen] >> 2);
         value.bits.ignored = 0;
         value.bits.upgradable = 0;
     }
