+++ Source/Utilities/UI/DisplayHelpers.mm
@@ -62,7 +62,7 @@ NSUInteger DOMNodeList$countByEnumeratingWithState$objects$count$(DOMNodeList *s
     else if (length > count)
         length = count;
         for (size_t i(0); i != length; ++i)
-            objects[i] = [self item:state->state++];
+            objects[i] = [self item:(unsigned int)state->state++];
             state->itemsPtr = objects;
             state->mutationsPtr = (unsigned long *) self;
             return length;
